+++ src/config/sagas.js
@@ -14,6 +14,7 @@ import languages from 'modules/languages/sagas';
 import mangroveData from 'modules/mangrove-data/sagas';
 import mangroveSpeciesData from 'modules/mangrove-species-data/sagas';
 import mangroveBiomassData from 'modules/mangrove-biomass-data/sagas';
+import mangroveHeightData from 'modules/mangrove-height-data/sagas';
 import ranking from 'modules/ranking/sagas';
 import mangroveProtectionData from 'modules/mangrove-protection-data/sagas';
 import mangroveRestorationData from 'modules/mangrove-restoration-data/sagas';
@@ -39,6 +40,7 @@ export default function* root() {
     fork(mangroveData),
     fork(mangroveSpeciesData),
     fork(mangroveBiomassData),
+    fork(mangroveHeightData),
     fork(ranking),
     fork(mangroveProtectionData),
     fork(mangroveInvestmentPotential),