+++ src/widget-components/mangrove-height/component.js
@@ -1,83 +1,111 @@
-import React, { useEffect, useMemo } from 'react';
-import Select from 'components/select';
-import ChartWidget from 'components/chart-widget';
-import sortBy from 'lodash/sortBy';
-import { format } from 'd3-format';
+import React, { useEffect } from "react";
+import PropTypes from 'prop-types';
 
-import config from './config';
+// components
+import Select from "components/select";
+import ChartWidget from "components/chart-widget";
 
-const numberFormat = format(',.2f');
+// utils
+import { format } from "d3-format";
+
+import config from "./config";
+
+const numberFormat = format(",.2f");
 
 const MangroveHeight = ({
   data: rawData,
+  isLoading,
+  metadata,
   isCollapsed = true,
-  slug, name,
+  slug,
+  name,
   currentLocation,
   addFilter,
   ui,
   setUi,
+  fetchMangroveHeightData,
   ...props
 }) => {
+  const { id } = currentLocation;
   const { year } = ui;
-  const years = useMemo(() => sortBy(rawData, ['date', 'desc'])
-    .filter(({ hmax_hist_m }) => hmax_hist_m)
-    .map(({ date }) => date?.split('-')[0]).reverse(),
-  [rawData]);
+  const heightCoverage = metadata?.avg_height[0]?.value;
+  const years = metadata?.year;
 
   useEffect(() => {
-    const yearSelected = years[0];
-    addFilter({
-      filter: {
-        id: 'height',
-        year: year || yearSelected,
-        area: 'maximum'
-      }
-    });
-    setUi({ id: 'height', value: { year: year || yearSelected } })
-  }, [year, addFilter]);
- 
+    if (!id || id === 1561) {
+      fetchMangroveHeightData();
+    } else {
+      fetchMangroveHeightData({ location_id: id });
+    }
+  }, [id, fetchMangroveHeightData]);
+
+  useEffect(() => {
+    if (!isLoading) {
+      addFilter({
+        filter: {
+          id: "height",
+          year: years?.[0],
+        },
+      });
+      setUi({
+        id: "height",
+        value: {
+          year: year || years?.[0],
+        },
+      });
+    }
+  }, [setUi, year, years, addFilter, isLoading]);
+
   if (!rawData || !year) {
     return null;
   }
 
-  const { chartData, chartConfig, heightCoverage, downloadData } = config.parse(rawData, year, years);
+  const { chartData, chartConfig, downloadData } = config.parse(
+    rawData,
+    year,
+    years,
+    heightCoverage
+  );
 
   const location = currentLocation?.name;
-  
-  const dateOptions = () => years.map(year => ({
-    label: year.toString(),
-    value: year.toString()
-  }));
-  
+
+  const dateOptions = () =>
+    years.map((year) => ({
+      label: year.toString(),
+      value: year,
+    }));
+
   const dateHandler = (value) => {
-    setUi({ id: 'height', value: { year: value } });
+    setUi({ id: "height", value: { year: value } });
     addFilter({
       filter: {
-        id: 'height',
-        year: value
-      }
+        id: "height",
+        year: value,
+      },
     });
   };
-  
+
   const dateSelector = (
     <Select
-    value={year}
-    options={dateOptions}
-    onChange={value => dateHandler(value)}
+      value={year}
+      options={dateOptions}
+      onChange={(value) => dateHandler(value)}
     />
-    );
-    
-    const sentence = (
-      <>
-      Mean mangrove <strong>maximum</strong> canopy height in <strong>{location}</strong> was
-      <strong> {numberFormat(heightCoverage)} m</strong> in <strong>{dateOptions.length > 1 ? dateSelector : year}</strong>.
+  );
+
+  const sentence = (
+    <>
+      Mean mangrove <strong>maximum</strong> canopy height in{" "}
+      <strong>{location}</strong> was
+      <strong> {numberFormat(heightCoverage)} m</strong> in{" "}
+      <strong>{dateOptions.length > 1 ? dateSelector : year}</strong>.
     </>
   );
   const widgetData = {
     data: chartData,
-    config: chartConfig
+    config: chartConfig,
   };
-  
+
   if (!chartData || !chartData.length) {
     return null;
   }
@@ -94,6 +122,43 @@ const MangroveHeight = ({
       {...props}
     />
   );
+};
+
+MangroveHeight.propTypes = {
+  data: PropTypes.shape({}),
+  isLoading: PropTypes.bool,
+  metadata: PropTypes.shape({
+    avg_height: PropTypes.arrayOf(PropTypes.shape({
+      year: PropTypes.number,
+      value: PropTypes.number
+    })),
+    location_id: PropTypes.string,
+    note: PropTypes.string,
+    units: PropTypes.shape({}),
+    year: PropTypes.arrayOf(PropTypes.number),
+  }),
+  isCollapsed: PropTypes.bool,
+  slug: PropTypes.string,
+  name: PropTypes.string,
+  currentLocation: PropTypes.shape({}),
+  ui: PropTypes.string,
+  addFilter: PropTypes.func,
+  setUi: PropTypes.func,
+  fetchMangroveHeightData: PropTypes.func,
+};
+
+MangroveHeight.defaultProps = {
+  data: null,
+  isLoading: true,
+  metadata: null,
+  isCollapsed: false,
+  slug: null,
+  name: null,
+  currentLocation: null,
+  ui: null,
+  addFilter: () => {},
+  setUi: () => {},
+  fetchMangroveHeightData: () => {}
 };
 
 export default MangroveHeight;